requirements.txt
altair==4.2.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.0
astor==0.8.1
asttokens==2.0.5
attrs==21.4.0
backcall==0.2.0
backports.zoneinfo==0.2.1
base58==2.1.1
black==22.1.0
bleach==4.1.0
blinker==1.4
blis==0.7.5
cachetools==5.0.0
catalogue==1.0.0
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.11
click==8.0.3
cryptography==36.0.1
cymem==2.0.6
dateparser==1.1.0
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
Django==4.0.2
djangorestframework==3.13.1
docx2txt==0.8
entrypoints==0.4
executing==0.8.2
gitdb==4.0.9
GitPython==3.1.26
idna==3.3
importlib-resources==5.4.0
ipykernel==6.8.0
ipython==8.0.1
ipython-genutils==0.2.0
ipywidgets==7.6.5
jedi==0.18.1
Jinja2==3.0.3
jsonschema==4.4.0
jupyter-client==7.1.2
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
langcodes==3.3.0
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
mistune==0.8.4
murmurhash==1.0.6
mypy-extensions==0.4.3
mysqlclient==2.1.0
nbclient==0.5.10
nbconvert==6.4.1
nbformat==5.1.3
nest-asyncio==1.5.4
notebook==6.4.8
numpy==1.22.2
packaging==21.3
pandas==1.4.0
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.9.0
pathy==0.6.1
pdfminer.six==20211012
pdfplumber==0.6.0
pexpect==4.8.0
phonenumbers==8.12.42
pickleshare==0.7.5
Pillow==9.0.1
plac==1.1.3
platformdirs==2.4.1
preshed==3.0.6
prometheus-client==0.13.1
prompt-toolkit==3.0.26
protobuf==3.19.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==7.0.0
pycparser==2.21
pydantic==1.8.2
pydeck==0.7.1
Pygments==2.11.2
Pympler==1.0.1
pyparsing==3.0.7
pyrsistent==0.18.1
python-dateutil==2.8.2
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
pyzmq==22.3.0
regex==2022.1.18
requests==2.27.1
Send2Trash==1.8.0
six==1.16.0
smart-open==5.2.1
smmap==5.0.0
spacy==2.3.5
spacy-legacy==3.0.8
spacy-loggers==1.0.1
sqlparse==0.4.2
srsly==1.0.5
stack-data==0.1.4
streamlit==1.5.0
terminado==0.13.1
testpath==0.5.0
thinc==7.4.5
tika==1.24
toml==0.10.2
tomli==2.0.0
toolz==0.11.2
tornado==6.1
tqdm==4.62.3
traitlets==5.1.1
typer==0.4.0
typing-extensions==4.0.1
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.8
validators==0.18.2
Wand==0.6.7
wasabi==0.9.0
watchdog==2.1.6
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.2
zipp==3.7.0
